+++ Changelog.md
@@ -4,6 +4,12 @@ RibbonExtensions:
 
 - New SetAvailabilityController() extension
 
+ParameterExtensions:
+
+- New Set(bool) extension
+
+Nuget symbol server support: https://symbols.nuget.org/download/symbols
+
 # Release 2023.1.6
 
 CollectorExtensions: