+++ Changelog.md
@@ -7,6 +7,7 @@ RibbonExtensions:
 ParameterExtensions:
 
 - New Set(bool) extension
+- New Set(color) extension
 
 Nuget symbol server support: https://symbols.nuget.org/download/symbols